==> prepocessing/s1-clean-eclipse-append-cmaq.ipynb <==
import pystac_client
import planetary_computer
import numpy as np
import pandas as pd
import geopandas as gpd
from datetime import timedelta,date,datetime
import matplotlib.pyplot as plt
import dask.dataframe as dd
#from netCDF4 import Dataset

# Order of operations:

1. **dataAnalysis/clean_eclipse_append_cmaq.ipynb : pull in eclipse, make hourly average, match CMAQ to hourly eclipse data**

2. dataAnalysis/clean_eclipse_find_nearest.ipynb : pull in output from #1, drop bad sensors

3. testML/format_input_data.ipynb : pull in output from #2, add in reanalysis data, ndvi, crop to chicago domain

4. dataAnalysis/format_input_shapefiles.ipynb: final form for monthly average data, add in zoning, put onto shapefile

# pull data using Internal API (to get the 5 min values)


import requests 
import json
import pandas as pd

def get_eclipse_data(devices: list, start_datetime_utc: str, end_datetime_utc: str) -> pd.DataFrame:
    # get data from url
    #devices = str(devices)[1:-1]
    eclipse_raw = []
    for device in devices:
        url = requests.get("https://eclipseprowebapi.azurewebsites.net/EclipseData/GetReadingsForCustomer?CustomerName=Chicago"
                            "&API_KEY=494FE543-82B7-45A2-8C1D-D9209AE45C94"
                            "&BeginDateTime={}&EndDateTime={}"
                            "&DeviceSubset={}".format(start_datetime_utc, end_datetime_utc, device))
        text = url.text
        data_dict_list = json.loads(text)
        if isinstance(data_dict_list, list):
            eclipse_raw += data_dict_list
    eclipse_raw = pd.DataFrame.from_records(eclipse_raw)
    return eclipse_raw

#devices3 = [2106, 2108, 2129, 2060, 2077, 2093, 2069]
alldevice = [2002 2003 2004 2005 2006 2008 2009 2010 2012 2013 2014 2017 2038 2037 2042 2043 2046 2048 2049 2050 2052 2053 2054 2055 2056 2058 2060 2062 2063 2064 2065 2066 2067 2068 2069 2070 2072 2073 2074 2076 2077 2079 2080 2081 2082 2083 2084 2090 2091 2092 2093 2094 2095 2096 2097 2098 2099 2100 2101 2102 2103 2104 2105 2106 2107 2108 2110 2111 2112 2113 2115 2116 2118 2119 2120 2121 2122 2123 2124 2125 2127 2128 2129 2130 2131 2132 2134 2135 2136 2138 2139 2142 2143 2144 2150 2162 2172 2174 2175 2176 2177 2179 2182 2183 2184 2185 2191 2207 2212 2047 2137 2208 2209]
raw = get_eclipse_data(alldevice,"2022-02-01","2021-03-01")
raw.to_csv('../calibrated_no2_full_data-Feb2022.csv')

# Read in data if already pulled

df = pd.read_csv('../calibrated_no2_full_data-Feb2022.csv',index_col = 0)
#df = pd.read_csv('calibrated_no2_full_data-Aug2021.csv',index_col = 0)

df['DeviceId'] = df['MSRDeviceNbr']
df['ReadingDateTimeUTC'] = pd.to_datetime(df.ReadingDateTimeUTC)
# drop zero values
df = df[df['CalibratedPM25']>0].reset_index(drop=True)
df = df[df['CalibratedNO2']>0].reset_index(drop=True)

df = df.sort_values('ReadingDateTimeUTC').reset_index(drop=True)

# Clean eclipse data

Resample for 1 hour measurements, drop when less that 75% of hourly data is available

#
st_dt ='2022-02-01'
end_dt ='2022-02-28 23:00:00'


# Clean data given the number of readings -- if less than 12 readings in an hour drop
df2 = df[['CalibratedPM25','ReadingDateTimeUTC','DeviceId']]
df2s = df2.groupby(["DeviceId"]).resample("H",on="ReadingDateTimeUTC").count()
under_count = (df2s.DeviceId >= 9) # use this as drop index

# drop given the restrictions
ts = df.groupby(["DeviceId"]).resample("H",on="ReadingDateTimeUTC").mean()[under_count]
#print(ts)

# reindex ts so that nan values fill up missing time steps
ts = ts.drop(columns='DeviceId').reset_index()
t_index = pd.DatetimeIndex(pd.date_range(start= st_dt, end= end_dt, freq='1h'))
ids = df.DeviceId.unique()

x,y = [],[] # for later use for cmaq
ts_nomissing = pd.DataFrame()
missing_ids = []
# create dataframe that contains every hour of data
for i in range(len(ids)):
  if len(ts[ts['DeviceId'] == ids[i]]) > 0:
    tmp = ts[ts['DeviceId'] == ids[i]].reset_index(drop=True)
    xx = tmp.Longitude[0]; yy = tmp.Latitude[0]
    x.append(xx); y.append(yy)
    # reindex with complete hourly spread
    tmp = tmp.set_index(pd.to_datetime(tmp['ReadingDateTimeUTC']))
    tmp = tmp.resample('1h').mean().reindex(t_index).fillna(np.nan)
    # Backfill static data
    tmp['DeviceId'] = [ids[i]]*len(tmp)
    tmp['Latitude'] = [yy]*len(tmp)
    tmp['Longitude'] = [xx]*len(tmp)
    tmp['latlon'] = [(xx,yy)]*len(tmp)
    ts_nomissing = ts_nomissing.append(tmp)
  else: missing_ids.append(ids[i])

ts_nomissing

drop_ids = []
# identify sensors with < 20 days of data
for i in range(len(ids)):
    tmp = ts_nomissing[ts_nomissing.DeviceId == ids[i]]        
    if np.sum(~np.isnan(tmp['CalibratedNO2'])) < (len(tmp)*.75):
        drop_ids.append(ids[i])


print(len(tmp)*(.75))
print(len(drop_ids))

for i in range(len(drop_ids)):
    ts_nomissing = ts_nomissing[ts_nomissing.DeviceId != drop_ids[i]].reset_index(drop=True)

ts_nomissing

ts_nomissing.to_csv('../dataAnalysis/cleaned_and_dropped_wint.csv')

# 

# Now append cmaq to eclipse data

Find corresponding indices that match eclipse lon/lat to cmaq lon/lat; pull data from cmaq; add these as columns in eclipse data

def find_index(stn_lon, stn_lat, wrf_lon, wrf_lat):
# stn -- list (points)
# wrf -- list (grid)
  xx=[];yy=[]
  for i in range(len(stn_lat)):
    abslat = np.abs(wrf_lat-stn_lat[i])
    abslon= np.abs(wrf_lon-stn_lon[i])
    c = np.maximum(abslon,abslat)
    latlon_idx = np.argmin(c)
    x, y = np.where(c == np.min(c))
    #add indices of nearest wrf point station
    xx.append(x);yy.append(y)
  #return indices list
  return xx, yy

from netCDF4 import Dataset

# Append CMAQ data
d = '../netcdf/'
lat,lon = np.array(Dataset(d+'/latlon_ChicagoLADCO_d03.nc')['lat']),np.array(Dataset(d+'/latlon_ChicagoLADCO_d03.nc')['lon'])

ds = Dataset('/mnt/batch/tasks/shared/LS_root/mounts/clusters/cmaq/code/Users/t-stacym/netcdf/all_crop_wint.nc')

#var_list = ['BENZENE','CO','FORM','NO', 'NO2', 'NOX', 'O3', 'PBLH',
# 'PM25_CA', 'PM25_CL', 'PM25_EC', 'PM25_HP', 'PM25_K', 'PM25_MG', 'PM25_NA', 'PM25_NH4', 'PM25_NO3', 
# 'PM25_OC', 'PM25_OM', 'PM25_SO4', 'PM25_SOIL', 'PM25_TOT',
# 'RH', 'SFC_TMP','SO2', 'U10', 'V10', 'VOC', 'precip']

# variables to add to dataframe
var_list = ['CO','FORM','NO', 'NO2', 'NOX', 'O3', 'PBLH',
 'PM25_EC', 'PM25_HP',  'PM25_NO3', 'PM25_OC', 'PM25_SOIL', 'PM25_TOT',
 'RH', 'SFC_TMP','SO2', 'U10', 'V10', 'VOC', 'precip']


ids = ts_nomissing.DeviceId.unique()

# find corresponding index in cmaq data for each station
x, y =[],[]
for i in ids:
  tmp = ts_nomissing[ts_nomissing.DeviceId == i].reset_index(drop=True)
  x.append(tmp['Longitude'][0])
  y.append(tmp.Latitude[0])


xx, yy = find_index(x,y,lon,lat) # get index of ID in cmaq file
txx,tyy = [],[]
#print(xx)

# get cmaq data to given index locations

# for each var
for var in var_list:
  cmaq = []
  # for each ID from eclipse
  for i in range(len(ids)):
    ind = xx[i][0],yy[i][0]
    # for each dat
    for d in range(28): # change this if doing august or february
      # for all hours
      for h in range(24):
        cmaq.append(ds[var][d*24+h][0][ind])
          if var == var_list[0]:
            txx.append(ind)
  ts_nomissing[var] = cmaq
  print('Done with var %s'%(var))



#reindexing arrays, keep it from look weird
f = ds[var][d*24:d*24+24].reshape(288,315,24,1)[ind]
f2 = [ds[var][d*24+h][0][ind] for h in range(24)]

#what does this look like in the end
ts_nomissing.Latitude

#ts_nomissing['xx'] = txx
# put out the file so that we can tmp use it
ts_nomissing.to_csv('cleaned_data_wint.csv')

# Now check nearby sensors to identify anticorrelated sensors

df = pd.read_csv('../cleaned_data_wint.csv',index_col = 0)
df['DeviceId'] = df['MSRDeviceNbr']
df2 = df.groupby('DeviceId').mean().reset_index()

ids = df.DeviceId.unique()

lo,la = [],[]

for id in ids:
    lo.append(df[df.DeviceId == id].Longitude.unique()[0])
    la.append(df[df.DeviceId == id].Latitude.unique()[0])


geometry = [Point(lo[i],la[i]) for i in range(len(lo))]

id = [ids[i] for i in range(len(ids))]
f = pd.DataFrame([id,lo,la]).T
f.columns = ['ids','lon','lat']

eclipse = gpd.GeoDataFrame(f,geometry = geometry,crs ={'init': 'epsg:4326'})
eclipse = eclipse.to_crs("EPSG:3857")

match = []
match_long = []
for i in range(len(ids)):
    match = []
    for j in range(len(eclipse)):
        point_1 = eclipse.geometry[i]
        point_2 = eclipse.geometry[j]
        d = point_1.distance(point_2)
        if (d < 5000) & (d != 0):
            match.append(int(eclipse.ids[j]))
            #print(match)
    match_long.append(match)

f = pd.DataFrame(match_long)
f['id'] = id
f.to_csv('nearest_station_names.csv')

data = pd.read_csv('../cleaned_data_wint.csv')
ids = data.DeviceId.unique()
loc = []

for id in ids:
    loc.append([data[data.DeviceId == id].Longitude.unique()[0],data[data.DeviceId == id].Latitude.unique()[0]])

import warnings 
warnings.filterwarnings('ignore')

df['dt'] = pd.to_datetime(df['Unnamed: 0'])

times = df.dt.unique()
times = pd.date_range(times[0],times[-1])
df['day_dt'] = [df['Unnamed: 0'][i][0:10] for i in range(len(df))]
df['day_dt'] = pd.to_datetime(df['day_dt'])

rs = []
for id in ids:
    tmp = df[df.DeviceId == id].reset_index()
    closest = np.array(f[f.id == id])[0]
    n = (np.sum(~np.isnan(closest))) # number of close sensors
    r = []
    if n > 1:
        for j in range(n-1):
            ri = []
            tmp_closest = df[df.DeviceId == closest[j]].reset_index()
            for i in range(len(times)):
                t = np.array(tmp[tmp.day_dt == times[i]]['CalibratedNO2'])
                t2 = np.array(tmp_closest[tmp_closest.day_dt == times[i]]['CalibratedNO2'])
                if (len(t2) > 0) & (len(t) > 0):
                    m1= np.array([~np.isnan(t),~np.isnan(t2)])
                    m1 = m1.all(0)
                    #print(m1)
                    t,t2 = t[m1],t2[m1]
                    ri.append(np.corrcoef(t,t2)[0][1])
            r.append(ri)    
        #print(np.array(r).shape)
        #print(r)
        rs.append(list(np.nanmean(np.array(r),axis=0)))
    else:
        rs.append([np.nan]*len(times))

pd.DataFrame(rs)

check = pd.DataFrame(
    [ids,
    list(pd.DataFrame(rs).T.mean()),
    list(pd.DataFrame(rs).T.min()),
    list(pd.DataFrame(rs).T.max())]
    ).T

check

rs = pd.DataFrame(rs)
rs['Correlated Days'] = np.array(rs > 0).sum(axis=1)-1
a = np.sort(rs[rs['Correlated Days'] < 31*0.75].id)
a
b = np.array([2118, 2080, 2134, 2060, 2082, 2062, 2079, 2048, 2172, 2182, 2179, 2191, 2162]) # form hourly concentraitons per month
np.sort(b)

s = list(a)+list(b)
s = pd.DataFrame([s,s]).T[0].unique()
print(len(s))
print(s)

locs = pd.DataFrame(loc)
locs['id'] = ids


fig,ax = plt.subplots(figsize=(9,12))
ax.scatter(locs[0],locs[1])

for i in range(len(s)):
    tmp = locs[locs.id == s[i]]
    ax.scatter(tmp[0],tmp[1])

plt.show()

# drop sensors bad
for i in range(len(s)):
    indexNames = data[data['DeviceId'] == s[i]].index
    data.drop(indexNames , inplace=True)

data.reset_index(drop=True)
data.to_csv('data_cleaned_and_dropped_wint.csv')

fin = pd.read_csv('data_cleaned_and_dropped.csv',index_col = 0)
fin = fin.groupby(['Latitude','Longitude']).mean().reset_index()
chi_rect_shp = gpd.read_filegrid = gpd.read_file('../shapefiles/rectangle_chicago_1km_grid.shp')
chi_shp = gpd.read_file('../shapefiles/geo_export_77af1a6a-f8ec-47f4-977c-40956cd94f97.shp')

# for plotting
streets = gpd.read_file('../shapefiles/geo_export_d268c11a-e3fe-42de-a88f-ad64688102f5.shp')
class1 = streets[streets['class'] == '1'].reset_index(drop=True)

chi_shp = chi_shp.to_crs("EPSG:4326")
outer = gpd.GeoDataFrame({'index':[1],'geometry':chi_rect_shp.unary_union},crs='EPSG:4326')
outer_chi = gpd.GeoDataFrame({'index':[1],'geometry':chi_shp.unary_union[0]},crs='EPSG:4326')

class1 = class1.to_crs(chi_shp.crs)


# make figure 1

import matplotlib.pyplot as plt

fig,ax = plt.subplots(figsize = (10,10))

outer.plot(facecolor='None',edgecolor='k',alpha=0.5,ax=ax,linewidth=1)
outer_chi.plot(facecolor='None',edgecolor='k',alpha=0.5,ax=ax,linewidth=1)

chi_shp.plot(facecolor='None',edgecolor='k',alpha=0.5,ax=ax,linewidth=0.5)
class1.plot(edgecolor='midnightblue',alpha=0.5,ax=ax,linewidth=1.5)

ctx.add_basemap(ax,crs=chi_shp.crs,source=ctx.providers.CartoDB.PositronNoLabels) #ctx.providers.Esri.NatGeoWorldMap)

ax.scatter(df2.Longitude, df2.Latitude,color='slategray',alpha=0.7,s=15,label='Excluded Station',marker='^')
ax.scatter(fin.Longitude, fin.Latitude,color='royalblue',s=20,label=r'NO$_2$ Sensor')


# add label
ax.text(0.10,0.96,s='(a)',transform=ax.transAxes,fontsize = 12,c='k',alpha=0.8)
ax.text(0.22,0.87,s='(b)',transform=ax.transAxes,fontsize = 12,alpha=0.8)

ax.text(0.14,0.795,s="O'Hare\n   ✈",transform=ax.transAxes,fontsize = 10,alpha=0.8)
ax.text(0.4,0.45,s='Midway\n   ✈',transform=ax.transAxes,fontsize = 10,alpha=0.8)
ax.text(0.45,0.14,s='EPA ComED',transform=ax.transAxes,fontsize = 10,alpha=0.8)

ax.text(0.74,0.35,s="I-90",transform=ax.transAxes,fontsize = 9,alpha=0.8, style='italic')
ax.text(0.53,0.6,s='I-290',transform=ax.transAxes,fontsize = 9,alpha=0.8, style='italic')
ax.text(0.53,0.48,s='I-55',transform=ax.transAxes,fontsize = 9,alpha=0.8, style='italic')
ax.text(0.45,0.75,s='I-94',transform=ax.transAxes,fontsize = 9,alpha=0.8, style='italic')
ax.text(0.65,0.35,s='I-94',transform=ax.transAxes,fontsize = 9,alpha=0.8, style='italic')
ax.text(0.58,0.2,s='I-57',transform=ax.transAxes,fontsize = 9,alpha=0.8, style='italic')
ax.text(0.75,0.2,s='I-94',transform=ax.transAxes,fontsize = 9,alpha=0.8, style='italic')
ax.text(0.72,0.58,s='LSD',transform=ax.transAxes,fontsize = 9,alpha=0.8, style='italic')

ax.text(0.8,0.7,s='  Lake\nMichigan',transform=ax.transAxes,fontsize = 12,alpha=0.5, style='italic')


plt.legend()

plt.show()

#ts_nomissing = ts_nomissing.reset_index(inplace=True)
err = [2048,2172,2182,2179,2191,2964,2062,2082,2079,2176,2069,2080]
#ts_nomissing.iloc(ts_nomissing.index([ts_nomissing['MSRDeviceNbr'] == err[0]]))['CalibratedNO2'] = np.nan
ts_nomissing
#ts_nomissing.index([ts_nomissing['MSRDeviceNbr'] == err[0]])

# drop errenous sensors
err = [2048,2172,2182,2179,2191,2964,2062,2082,2079,2176,2069,2080]
for e in err:
    ts_nomissing.at([ts_nomissing['MSRDeviceNbr'] == e]['CalibratedNO2']) = np.nan


ts_nomissing.to_csv('cleaned_data_wint.csv')

import geopandas as gpd
from shapely.geometry import Point, shape, Polygon

# Plot the monthly averages

monthly_avg = ts_nomissing.groupby('DeviceId').resample('M').mean().reset_index(drop=True)
monthly_avg['geometry'] = [Point(monthly_avg.Longitude[i],monthly_avg.Latitude[i]) for i in range(len(monthly_avg))]
monthly_avg = gpd.GeoDataFrame(monthly_avg)
monthly_avg=monthly_avg.set_crs("EPSG:4326")

ts_nomissing['CalibratedPM25']

from shapely.geometry import mapping
import rioxarray as rxr
import xarray as xr
import geopandas as gpd

f = gpd.read_file('shapefiles/tmp_clipped.shp')
f = f.to_crs(monthly_avg.crs)
hwy = gpd.read_file('shapefiles/hwy.shp')
hwy = hwy.to_crs(monthly_avg.crs)

fig,ax = plt.subplots(figsize=(10,10))
f.plot('pm25',ax=ax,vmin=10,vmax=15)
hwy.plot(color='k',ax=ax,linewidth=0.25)
monthly_avg.plot('CalibratedPM25',vmin=10,vmax=15,ax=ax,s=20,edgecolor='white',linewidth=0.25,zorder=10,
                legend=True,legend_kwds={'fraction':0.03, 'pad':0.04, 'label':r'${\mu}g/m^3$'})
ax.axis('off')
ax.set_title(r'Monthly Average PM$_{2.5}$ over Chicago')
plt.show()

fig,ax = plt.subplots(figsize=(10,10))
f.plot('no2',ax=ax,vmin=10,vmax=20)
hwy.plot(color='k',ax=ax,linewidth=0.25)
monthly_avg.plot('CalibratedNO2',vmin=10,vmax=20,ax=ax,s=20,edgecolor='white',linewidth=0.25,zorder=10,
                legend=True,legend_kwds={'fraction':0.03, 'pad':0.04, 'label':r'${\mu}g/m^3$'})
ax.axis('off')
ax.set_title(r'Monthly Average NO2$_{2.5}$ over Chicago')
plt.show()

ts_nomissing.to_csv('cleaned_data.csv')


print(df[df['LocationName'] == 'EPA COM ED Maintenance Bldg D']['DeviceId'].unique()) # 2008
print(df[df['LocationName'] == 'EPA COM ED Maintenance Bldg C']['DeviceId'].unique()) #2017
print(df[df['LocationName'] == 'EPA COM ED Maintenance Bldg E']['DeviceId'].unique()) #2208

print(df[df['LocationName'] == 'EPA Springfield Pump Station C']['DeviceId'].unique()) #2010
print(df[df['LocationName'] == 'EPA Springfield Pump Station B']['DeviceId'].unique()) #2009
print(df[df['LocationName'] == 'EPA Springfield Pump Station E']['DeviceId'].unique()) #2212

print(df[df['LocationName'] == 'EPA Village Garage D']['DeviceId'].unique()) #2144
print(df[df['LocationName'] == 'EPA Village Garage C']['DeviceId'].unique()) #2014
print(df[df['LocationName'] == 'EPA Village Garage B']['DeviceId'].unique()) #2013


df[df['LocationName'] == 'EPA COM ED Maintenance Bldg D']
df[df.DeviceId == missing_ids[12]]


for i in range(len(missing_ids)):
    print(df[df.DeviceId == missing_ids[i]]['LocationName'].iloc[0])
