==> prepocessing/s4-download_5p_directly.ipynb <==
from sentinelsat import SentinelAPI, read_geojson, geojson_to_wkt
from datetime import date, timedelta, datetime 
import xarray as xr
import numpy as np
import fsspec

PRODUCT_TYPE = 'L2__NO2___'
START_DATE = date(2022,2,1)
END_DATE = date(2022,3,1) 
print("START DATE: ", START_DATE, ". END DATE: ", END_DATE)
#area = "POLYGON((-87.5171 41.6298,-87.4896 41.6934,-87.6736 42.0371,-87.9510 42.0269,-87.9648 41.9513,-87.7505 41.6647,-87.5171 41.6298)))"
dc = '../shapefiles/chicago_map_simple_final.geojson'
area = geojson_to_wkt(read_geojson(dc))


api = SentinelAPI('s5pguest', 's5pguest', 'https://s5phub.copernicus.eu/dhus/')
products = api.query(area, date = (START_DATE, END_DATE),platformname='Sentinel-5 Precursor', producttype=PRODUCT_TYPE )
df = api.to_dataframe(products)
df = df[df['processingmodeabbreviation'] == 'OFFL']
for i in df['link'] :
    print(i);

print(df)
print(len(df))
print(len(products))
# download all results from the search
#api.download_all(products)


import pandas as pd
pd.DataFrame(np.array(df.title)).to_csv('satnames_022022.csv')
